pyproject.toml
[project]
name = "ecommerce-core"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "pydantic-settings==2.4.0",
    "psycopg[binary]==3.2.10",
    "redis==5.0.0",
    "python-json-logger==2.0.7",
    "sqlalchemy==2.0.30",
    "alembic==1.13.2",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "python-jose[cryptography]==3.3.0",
    "email-validator==2.2.0",
    "python-multipart==0.0.9",
]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["alembic/versions", "alembic/README"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
